add endpoints for step 2

+++ backend/api/models/CoutRex.py
@@ -5,7 +5,7 @@ class CoutRex(BaseModel):
     num: int
     code_solution: int
     code_rex: int
-    cout_reel: int | None = None
+    cout_reel: float | None = None
     monnaie: str | None = None
     code_unite_cout: int | None = None
     code_difficulte: int | None = None
